Refactor: modularize game preview components, consolidate room polling utility, and implement dynamic routing via Games.jsx
